+++ R/bootcoldist.R
@@ -7,7 +7,7 @@
 #'  Can be the result from \code{vismodel}, or \code{colspace}. Data may also be
 #'  independently calculated quantum catches, in the form of a data frame with
 #'  columns representing photoreceptors.
-#' @param by (required) a vector containing indicating the group to which each row from
+#' @param by (required) a numeric or character vector indicating the group to which each row from
 #'  the object belongs to.
 #' @param boot.n number of bootstrap replicates (defaults to 1000)
 #' @param alpha the confidence level for the confidence intervals (defaults to 0.95)
@@ -24,7 +24,7 @@
 #'  data(sicalis)
 #'  vm <- vismodel(sicalis, achro='bt.dc')
 #'  gr <- gsub("ind..", "", rownames(vm))
-#'  bootcoldist(vm, gr, n = c(1, 2, 2, 4), weber = 0.1, weber.achro = 0.1, cores = 1)
+#'  bootcoldist(vm, by = gr, n = c(1, 2, 2, 4), weber = 0.1, weber.achro = 0.1, cores = 1)
 #'  }
 #'
 #' @export
@@ -103,8 +103,6 @@ bootcoldist <- function(vismodeldata, by, boot.n = 1000, alpha = 0.95,
   if (is.null(arg0$weber.ref)) {
     arg0$weber.ref <- "longest"
   }
-
-
 
   sortinggroups <- order(by)
   vismodeldata <- vismodeldata[sortinggroups, ]